+++ R/vgamParam.R
@@ -1,4 +1,4 @@
-#' Run a Generalized Additive Model on all voxels of a NIfTI image within a mask and and return parametric and smooth coefficients tables
+#' Run a Generalized Additive Model on all voxels of a NIfTI image within a mask and return parametric and smooth coefficients tables
 #'
 #' This function is able to run a Generalized Additive Model (GAM) using the mgcv package. 
 #' The analysis will run in all voxels in in the mask and will return parametric and smooth coefficients. 
@@ -13,7 +13,7 @@
 #' @param ... Additional arguments passed to gam()
 #' 
 #' @return Return list of parametric and spline coefficients (include standard errors and p-values) fitted to each voxel over the masked images passed to function.
-#' 
+#' @keywords internal
 #' @export
 #' @examples
 #' image <- oro.nifti::nifti(img = array(1:1600, dim =c(4,4,4,25)))
@@ -37,7 +37,7 @@ vgamParam <- function(image, mask , fourdOut = NULL, formula, subjData, mc.presc
   if (class(image) == "character" & length(image) == 1) {
     image <- oro.nifti::readNIfTI(fname=image)
   } else if (class(image) == "character" & length(image) > 1) {
-    image <- mergeNiftis(inputPaths = image, direction = "t", outfile <- fourdOut)
+    image <- mergeNiftis(inputPaths = image, direction = "t", outfile = fourdOut)
   }
   
   if (class(mask) == "character" & length(mask) == 1) {