+++ NEWS.md
@@ -3,6 +3,22 @@
 Functions for the voxelwise analysis of NIfTI data in R.
 
 #### Change log
+
+v1.3.2 2017.07
+
+* New function to run fsl randomise on generalized additive models
+* Clean documentation hid redundant functions
+* The functions have a flag to output residual maps (for cluster correction)
+
+v1.3.1 2017.07
+
+* Not released
+
+v1.2.2 2017.03
+
+* Minor edits to documentation
+* Not released
+
 v1.2.1 2016.08
 
 * First release